fix(autoscaling): can't use `MathExpression` in `scaleOnMetric`
It was not possible to use `MathExpression` objects in AutoScaling,
because the `Alarm` interface has a bug: it takes "override" arguments
like "period" and "statistic" that could and should have already been
encoded into the `Metric` object passed to the alarm.

Subsequently, AutoScaling used this ill-advised feature to force the
period of metrics to 1 minute. This caused an invalid render for math
expressions and was probably ill-advised to begin with. We should be
respecting the customer's period on the metric they pass in.

Fixes #5776.

BREAKING CHANGE: AutoScaling by using `scaleOnMetric` will no longer
force the alarm period to 1 minute, but use the period from the Metric
object instead (5 minutes by default). Use `metric.with({ period:
Duration.minute(1) })` to create a high-frequency scaling policy.