fix(ecs): cannot separate `Cluster` and `Ec2Service` behind ALB
In the case of an ECS Service runningon  EC2 capacity, pointing a Load
Balancer to the Service involves updating the SecurityGroups of the
capacity associated to the Cluster (to allow traffic from the Load
Balancer).

If these resources are in different stacks, this is liable to create
cyclic references: Service points to Cluster, Security Groups point
to Load Balancer, cyclic references arise from the way these resources
are typically colocated.

This changes makes it so the ingress/egress rules will be created in
the same stack as the Service (which points to both Cluster and LB),
hence making sure there are no cycles.