fix(apigateway): LambdaRestApi fails when a user defined Stage is attached (#5838)
LambdaRestApi, by default, sets up a proxy integration as part of its
construction which in turn binds the specified lambda function handler
to the proxy using the LambdaIntegration class.

The LambdaIntegration class needs to set up the right IAM permissions
so API Gateway is able to invoke the Lambda function. The policy for
this permission requires the ARN of the 'ANY' Method that is set up by
the Proxy.

When the 'deploy' option is set, it indicates to the RestApi construct
and its sub-constructs (of which LambdaRestApi is one) to create a
default deployment and stage as part of its construction. The user can
also unset the 'deploy' option and specify their own Stage by setting
the 'deploymentStage' on the RestApi construct.

However, when the 'deploy' option is unset, the LambdaIntegration class
cannot compute the ARN for the 'ANY' Method since there's no Stage and
Deployment created. This generates at synthesis, even though there may
be a Stage that the user has configured against the RestApi later on.

The fix here is straightforward. Computation of the Method ARN should
occur at the time of synthesis rather than at the time of construction.

fixes #5744

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>