feat(lambda-destinations): option to auto-extract the payload when using LambdaDestination (#5503)
* feat(lambda): asynchronous lambda chain

Add a `LambdaChain` construct to chain asynchronous invocations of Lambda
functions.

Asynchronously invoking the first function starts the execution chain. The
response payload of a successful invocation of a function in the chain is
passed as request payload to the next function automatically using
EventBridge.

Allows to start a series of long running processes.

The poor man's step function?

* LambdaPayloadDestination

* responseOnly option

* document default for responseOnly

* README and comment

* improve and document integ test

* JSDoc

* Apply suggestions from code review

Co-Authored-By: Niranjan Jayakar <16217941+nija-at@users.noreply.github.com>

* README and JSDoc

Co-authored-by: Niranjan Jayakar <16217941+nija-at@users.noreply.github.com>