feat(lambda): avail function log group in the CDK
The lambda function's log group is now available for use in the CDK app
so it can be used further to create subscription filters, metric
filters, etc. or in any other way that a regular log group may be used.

The implementation uses an underlying CustomResource to guarantee the
existence of this log group as part of the CloudFormation stack.
A new property called `exposeLogGroup` to enable this, so that existing
stacks that have a Lambda function don't change significantly by
automatically getting this CustomResource, as well as, users who are
not interested in this log group can opt out.

closes #3838