feat(eks): fargate profiles (#5589)
Add support for Fargate profiles. Profiles can be added to a cluster through the `addFargateProfile` method or by defining a `FargateProfile` construct.

The EKS custom resource was extended to support a new resource type that provisions Fargate profiles and slightly refactored to allow maximum code reuse between the two resource types.

The IAM policy for the cluster role was extended to allow creation of Fargate Profiles (`eks:CreateFargateProfile`, `eks:DescribeFargateProfile`, `eks:DeleteFargateProfile`, `iam:GetRole`, and `iam:CreateServiceLinkedRole`). Since we don't know the name of the profile/cluster during synthesis, the IAM policy is wider than desired (uses `*` for resource ARNs).

In order to enable Fargate-only cluster, the CoreDNS inside the Kubernetes cluster must be patched to use "fargate" as the compute type for CoreDNS. To that end, an additional resource, `KubernetesPatch`  was added to the kubctl custom resource provider. This resource type takes "apply" and "restore" manifests which are patched upon CREATE and DELETE, respectively.

The `FargateCluster` high level construct is a sugar for defining a Fargate-only EKS cluster.

Resolves #5303 

BREAKING CHANGE: (experimental module) the `Mapping` struct was renamed to `AwsAuthMapping`.