fix(autoscaling): can't use block devices
Block Devices for AutoScalingGroups were broken because they would
always synthesize `NoDevice: true|false`, but the `NoDevice` key should
not be present at all if it is not intended. In essence, `NoDevice:
false` should never occur.

Add a feature to the `@aws-cdk/assert` library to assert absence
of certain keys (by use of a magic value).

Fixes #5868.