fix(ec2): add `MachineImage` factory, document instance replacement
Some MachineImages get updated automatically (causing unexpected
instance replacement), and some don't. Update the documentation to very
clearly note which image is which, and that using a particular type of
image may cause your instance to be replaced.

Also add a class with static factory functions to make an enum-like
pattern, to be in line with other places where the user has a choice
of implementations.

Fixes #5675.

Also note that VPC lookups are cached, which hopefully addresses #6025.