fix(aws-stepfunctions): enable task property 'heartbeat'
A Task instance's 'taskProps' property adhereing to the StepFunctionsTaskConfig
interface. This interface currently supports the 'heartbeat' optional property.
The Task instance constructor takes in a props object adhering to the TaskProps
interface. The TaskProps interface does not currently support the 'heartbeat'
optional property.  The props object is then assigned to the Task instance's
'taskProps' property. By adding the 'heartbeat' optional property to the
TaskProps interface, we are enabling the 'heartbeat' optional property to be set
on the Task instance's 'taskProps' property.