chore: monocdk experiment (#5990)
An experiment to bundle the entire CDK into a single monolithic module called monocdk-experiment.

This module depends on all other modules as a dev-dependency and during build copies all the lib/** files
from all modules into an src/ directory and rewrites the "import" statements for internal references.

The goal is to release this to npm as an experiment in monolithic packaging.