chore(cognito): switch to jest
Migrate all existing unit tests from nodeunit to jest.

As part of this, centralize jest configuration in the `cdk-build-tools`
package with the ability for individual packages to override these via
the `package.json`.

Dropping the `JestCoverageTarget` pkglint rule since this change
ensures that a standard set of rules, that include coverage target,
will be applied to any package that is configured to use `jest`. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

<!-- 
Please read the contribution guidelines and follow the pull-request checklist:
https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md
 -->