fix(ecs-patterns): allow imported load balancers as inputs
This PR adds functionality in aws-ecs-patterns to allow imported Network and Application load balancers to be used when initializing the LoadBalancedEC2Service and LoadBalancedFargateService constructs. By necessity, this means:

1) editing the Network and Application load balancer constructs to add an optional IVpc property to be used when calling the from<LBType>LoadBalancerAttributes static methods, and 
2) changing the error which is thrown when calling addTargets on imported load balancers. 

Unit tests are added in aws-ecs-patterns to test the new import functionality and in aws-elasticloadbalancingv2 to ensure that addTargets behaves properly when a Vpc is or is not specified for imported load balancers. 

Resolves: https://github.com/aws/aws-cdk/issues/5209