feat(cx-api): clean up features.ts (#6181)
Clean up features.ts such that it will include only things related to feature flags in order to make it easier for users to discover and add feature flags in one place. Merged it with the contents of `future.ts` and updates contribution guide to that end.

The other context keys that were in `feature.ts` were not actually feature flags but rather application switches, so these were moved to `app.ts`.

Fixes #6098