chore: make VPC-import integ test isolated (#6278)
The "VPC import" CLI integ test creates a VPC with a tag of `custom-tag:
bazinga!` and then tries to import it.

If multiple instances of this test run at the same time, multiple VPCs
will be created and the lookup will fail since the query must produce
a unique VPC. What's worse, the VPC won't even be cleaned up anymore
and all future test runs will fail.

Make the tag value include a unique token so that the query will
always look up the VPC just created. Also try to recover and delete
the stack we just created if the test fails.