chore: automatically approve dependabot PRs
Right now it falls to @NLG321 to approve all Dependabot PRs. When he
gets around to it, it's likely that conflicts have been introduced into
the `package.json` and `package-lock.json` files and the PR won't
automatically merge anymore.

Let's take the human out of the loop and use GitHub Actions to
automatically approve Dependabot's PRs, and configure Mergify
appropriately for Dependabot PRs.