feat(appsync): more general mapping template for DynamoDB PutItem (#6236)
* feat(appsync): more general mapping template for DynamoDB PutItem

Currently, the mapping template for DynamoDB PutItem doesn't support
specifying a full primary key, and it doesn't support picking values to
assign to attributes.

This patch introduces a more general approach, which allows setting a
full primary key as well as picking which values from the mutation
argument to assign to table attributes.

It keeps the previous convenient approach of projecting an object to a
set of attribute value assignments via the `Values.projecting` function,
which can then be augmented with more attributes. Example:

```ts
// saveCustomerWithFirstOrder(
//        customer: SaveCustomerInput!,
//        order: FirstOrderInput!,
//        referral: String): Order

MappingTemplate.dynamoDbPutItem(
    PrimaryKey
        .partition('order').auto()
        .sort('customer').is('customer.id'),
    Values
        .projecting('order')
        .attribute('referral').is('referral'));

```

BREAKING CHANGE: Changes `MappingTemplate.dynamoDbPutItem()` to accept
`PrimaryKey` and `AttributeValues`, which allow configuring the primary
key and to project an object to a set of attribute values.

Fixes #6225

Signed-off-by: Duarte Nunes <duarte@uma.ni>

* test(integ.graphql.ts): add test case for dynamoDbPutItem()

Signed-off-by: Duarte Nunes <duarte@uma.ni>

Co-authored-by: Mitchell Valine <mitchellvaline@gmail.com>