feat(kms): `trustAccountIdentities` avoids cyclic stack dependencies
Add a switch which will trust the entire account to use the key, subject to the identity policies of the identities making the calls (same as for all other AWS resources).

Since there is no way of updating a key policy after creating the key, AND bidirectional trust needs to be established between the resource and the identity, users are basically locked into creating both key and identities using the key in the same stack. The only way for cross-stack key usage where the key gets created first is by wildcarding. This switch allows that in an easy manner without introducing cyclic dependencies.