fix(core): top-level resources cannot use long logical ids
When allocating a logical ID for a CloudFormation resource we normally render the id by taking in the full construct path, trim it down to 240 characters and append a hash to it to make sure its unique. This technically allows construct IDs to be of any length.

However, when a resource is defined as a top-level resource (i.e. the child of a stack), it won't get this treatment and we will just use the construct id as it's logical id.

In this case, synthesis will fail with an error indicating the logical ID is invalid (too long).

This might work ok if the logical ID is human provided, but in some cases (like #6190), the id is actually created automatically and in some cases can be longer than 255.

This fix will only apply the special treatment to top-level resources if their construct ID is shorter than the maximum allowed length (255). Otherwise, it will use the same mechanism to trim them and append the hash.

Fixes #6190.