chore: master version in one place (#6410)
* chore: set lerna version 999.999.999

To avoid merge conflicts when merging back a release, update the checked-in repo version number to 999.999.999.

Also, use fixed versions across the board (no carets!).

* chore: store version only in root package.json

Store the current repo version only in the root pacakge.json in order to avoid the need to specify the version number everywhere in the codebase.

In order to achieve that, I changed all package.json files to use a version marker 999.999.999 to indicate it's a local dependency. This marker is better than 0.0.0.0 since it will make sure cache is invalidated for various languages (for example, Maven will only bring in a new version if the local cache has an older version).

Then, we needed to change a couple of other things:

- The bump script is now very simple. It merely updates the root package.json and creates the CHANGELOG. That's it.
- Before running a CI build (buildspec.yml), we run `scripts/set-version.sh` which uses lerna to update the version in all package.json files. This means that build artifacts that are produced in CI builds get the actual repo version, and dev builds all use 999.999.999.

* Use 999.0.0 instead of 999.999.999

* Move all configuration to bump.sh

Move configuration from package.json to bump.sh and change the commit title to `chore(release): vX.Y.Z`

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>