chore(monocdk): @monocdk-experiment/assert (#6362)
* chore(monocdk): @monocdk-experiment/assert

Create a clone (during build) of @aws-cdk/assert that takes a dependency on monocdk-experiment instead of @aws-cdk/core, so we can use it when playing around with mono-cdk.

* remove jest.ts (we are going to copy it as well)

* Remove dependency between cloudformation-diff and cx-api

In order to avoid the need to duplicate cfn-diff to work against monocdk-experiment.

* Copy jest.ts

* Import rewrite-imports from rix0rrr’s repo

* Use rewrite-imports

* Do not rewrite imports in files not listed

* Change import statements to be compatible with import rewriter

* Support disabling linters

* Allow monocdk-experiment to be used locally

* Disable linters when compiling @monocdk-experiment/assert

* use peer dependency

* cleaned up tsconfig.json

* Move rewrite-imports to bin/

* some CR updates

* fix foreach is "scripts" is not defined in package.json

* Update package.json

* Update package.json

* improve version alignment script

instead of using lerna, just written a simple node program that finds the version marker 999.0.0 and sets the version across the board.