chore: master version in lerna.json (instead of root package.json) (#6445)
* chore: master version in lerna.json (instead of root package.json)

Use lerna.json to master the repo version instead of the root package.json.

Move standard-version configuration to .versionrc.json and consult it when aligning the repo version to determine where the version is mastered.

* scripts/get-version

returns the current repo version, and reused by align-version. then we can use it in our AutoBump build project
instead of hard-coding the fact that lerna.json is where we master the version.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>