chore: use 0.0.0 as version marker & additional protection in pack.sh
Replace `999.0.0` with `0.0.0` to reduce the risk in accidental publishing. If, by accident, we will publish a module with the version `999.0.0`, it will semantically always be the latest version, and that’s catastrophic. By using `0.0.0` we reduce this risk. Worse case we will have a `0.0.0` module published, which is not a huge problem.

Codify this in `scripts/get-version-marker` and reuse everywhere.

We also improved the `pack.sh` script in the following ways:
- Use `get-version.js` to obtain the version for build.json (instead of reading from lerna.json)
- Fail if the version does not start with “1.” to protect against publishing a different major version
- Adds an additional check to verifies that none of the file names under `dist/` uses the “0.0.0” marker.