docs(apigateway): fixes typescript usage plan example (#6411)
Adds overall `throttle` setting to the example documentation on typescript CDK about how to use "an API Key with a usage plan". Without it, when trying to deploy the example code, it gives the following error:

```
Can not define fine grain throttling if usage plan throttle is not defined.
```

Before you can set individual throttle for each individual method, you need first to set the global throttle settings for that API. So adding the `throttle` parameter to the `addUsagePlan` method, with the same value as the individual throttle for the echoMethod, fixes it.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>