docs(apigateway): update tsc custom domain examples (#6413)
Explicitly add the necessary imports required for the route53 alias records examples in the custom domain section of the Typescript API Gateway documentation. Also, update the examples to use the same notation used in the route53 module, that is, `targets.` instead of `route53_targets.`

The previous examples were a bit confusing because they did not make it clear which packages were required, and even made it seem like you needed to import the targets from the route53 package. As an user who is learning the CDK and was not familiar with the Route53 and Route53 Targets packages I was stuck until I explored the documentation a bit more. I think this makes the example way more clear.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>