fix(cli): fast "no-op" deploys do not consider tags (#6472)
* fix(cli): fast "no-op" deploys do not consider tags

Stack-level tags apply to all supported resources in a stack. If they change, a
fast "no-op" deploy should not occur (e.g. a tag with the CDK version or a tag
with the commit hash when running in CI).

Fix the `'deploy not skipped if template changed'` test that was including the
`force` option.

Fix CLI integ tests not working anymore after #6463 and the `0.0.0` version.

* else

* compareTags

* README

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>