chore(pkglint): update language for experimental constructs.
The current language states that the entire module is unstable and
should not be used in production.

This is not entirely correct. The updated language differentiates CFN
resources from higher level APIs, and states the former can be used in
production and are stable, while the latter APIs are not stable and
subject to change.