fix(rds): setting timezone on DatabaseInstance causes internal failure (#6534)
* fix(rds): setting timezone on DatabaseInstance causes internal failure

Per documentation[1], 'Timezone' property is only supported on Microsoft
SQL Server. Setting this property on a DatabaseInstance with a different
database engine causes deployment to fail (1) with a validation error
for a new instance of `AWS::RDS::DBInstance` and (2) internal failure
when modifying an existing instance of `AWS::RDS::DBInstance`.

[1]:https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-rds-database-instance.html#cfn-rds-dbinstance-timezone

fixes #6439

* PR feedback