chore(pkglint): update language for experimental constructs. (#6500)
* chore(pkglint): update language for experimental constructs.

The current language states that the entire module is unstable and
should not be used in production.

This is not entirely correct. The updated language differentiates CFN
resources from higher level APIs, and states the former can be used in
production and are stable, while the latter APIs are not stable and
subject to change.

* add a sentence around intent

* tweak the L1 line

Co-Authored-By: Elad Ben-Israel <benisrae@amazon.com>

* slight tweak

* run pkglint -f and fixed up create-missing-libraries.ts

* revert unintended monocdk changes

Co-authored-by: Elad Ben-Israel <benisrae@amazon.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>