chore(cli): CLI uses `cdk-assets` to upload templates and assets
Centralize all logic about how templates, files and container assets
are built and uploaded in the `cdk-assets` tool.

We need this change first because it's on the critical path of the CLI
being able to deploy using the new convention mode roles (which requires
the CLI to use the asset publishing role for uploading the
CloudFormation templates), which is required for the pipeline
being able to do self-mutation using `cdk deploy`.

We can roll this change out independently of the framework emitting
the asset manifest (the CLI can generate it) and we don't need
to assume any roles, while we still get to test the new code path.

Also importing a number of improvements to `cdk-assets` from the
proof-of-concept branch.